+++ src/components/Layout/EmptyLayout.tsx
@@ -1,0 +1,7 @@
+import { LayoutPropsInterface } from '@/types/common';
+
+function EmptyLayout({ children }: LayoutPropsInterface) {
+  return <>{children}</>;
+}
+
+export default EmptyLayout;