+++ src/pages/chore/index.tsx
@@ -3,6 +3,11 @@ import Image from 'next/image';
 import lazyImage from '@/assets/images/lazy-chore-page.jpeg';
 import { APP_CONFIGS } from '@/configs/app';
 
+export async function getServerSideProps(){
+  return {
+    props: {},
+  }
+}
 function ChorePage() {
   return (
     <div className="text-center p-4">