+++ src/pages/confide/index.tsx
@@ -1,21 +1,14 @@
-import { useEffect } from 'react';
-import { getArticleList } from '@/services/api/getArticleList';
 import MainLayout from '@/components/Layout/MainLayout';
 
-// export async function getServerSideProps() {
-//   const resp = await getArticleList();
+export async function getServerSideProps() {
 
-//   const data = JSON.stringify(resp);
-//   console.log("data", resp)
-//   return {
-//     props: {
-//       data
-//     }
-//   }
-// }
+  return {
+    props: {}
+  }
+}
 
 function HomePage(props: any) {
-  return <div className="flex ">confide index</div>;
+  return <div className="flex ">confide page</div>;
 }
 
 HomePage.Layout = MainLayout;